requirements.txt
Flask==3.0.3
gunicorn==23.0.0
requests==2.32.5
python-dotenv==1.0.1
psycopg2-binary==2.9.11
boto3==1.34.162
PyPDF2==3.0.1
redis==5.0.8
sentry-sdk[flask]==2.13.0
backoff==2.2.1
pyjwt==2.9.0
Werkzeug==3.1.3
fastapi==0.115.6
uvicorn==0.34.0
python-dateutil==2.9.0.post0

